requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
backcall==0.2.0
braceexpand==0.1.7
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.6.0
entrypoints==0.4
executing==1.0.0
filelock==3.8.0
fonttools==4.37.1
frozendict==2.3.4
frozenlist==1.3.1
fsspec==2022.8.2
g2p-en==2.1.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
huggingface-hub==0.9.1
hydra-core==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.2.3
inflect==6.0.0
ipykernel==6.15.2
ipython==8.4.0
ipywidgets==8.0.1
jedi==0.18.1
joblib==1.1.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-widgets==3.0.2
kaldi-python-io==1.2.2
kaldiio==2.17.2
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.17.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.2
nemo-toolkit==1.10.0
nest-asyncio==1.5.5
nltk==3.7
numba==0.56.2
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.1.2
onnx==1.12.0
packaging==21.3
pandas==1.4.4
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pooch==1.6.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.1
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.13.0
pynini==2.1.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.6.4
pytz==2022.2.1
PyYAML==5.4.1
pyzmq==23.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.0
rich==12.5.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.1
sentencepiece==0.1.97
sentry-sdk==1.9.7
setproctitle==1.3.2
shellingham==1.5.0
shortuuid==1.0.9
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
sox==1.4.1
stack-data==0.5.0
sympy==1.11.1
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torchmetrics==0.9.3
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.2
typer==0.6.1
typing_extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.12
wandb==0.13.2
wcwidth==0.2.5
webdataset==0.1.62
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.2
wrapt==1.14.1
yarl==1.8.1
youtokentome==1.0.6
zipp==3.8.1
